ai-service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
mysql-connector-python>=8.0.0
python-multipart>=0.0.6
hypothesis>=6.90.0
pytest>=7.4.0
